feature(error): implicit error handling
All errors are implicitly shown using toast. It is up to the creater of
a promise to properly deal with any side-effects of a failed promise
(for example, properly setting variables to undefined or fallback
values). If nothing needs to be done, but just an error needs to be
logged, the `noop` function can be used to properly catch.